Finished the lookup controller (CRUD) for the EventCategories, polish…

+++ Persistence/DataContextInitializer.cs
@@ -5,6 +5,7 @@ using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
+using System.Diagnostics;
 using System.Drawing;
 
 namespace Persistence
@@ -39,8 +40,11 @@ namespace Persistence
             {
                 if (_context.Database.IsSqlite() || _context.Database.IsNpgsql())
                 {
-                    await _context.Database.EnsureDeletedAsync();
-                    await _context.Database.EnsureCreatedAsync();
+                    if (_context.Database.GetPendingMigrations().Count() > 0)
+                    {
+                        await _context.Database.EnsureDeletedAsync();
+                        await _context.Database.EnsureCreatedAsync();
+                    }
                     //await _context.Database.MigrateAsync();
                 }
             }
@@ -73,8 +77,8 @@ namespace Persistence
             var users = new List<User>
             {
                 new User { DisplayName = "Admin", UserName = "Admin", Email = "admin@test.com" },
-                new User { DisplayName = "Hephaestots", UserName = "vince", Email = "vinc@test.com" },
-                new User { DisplayName = "SilveR", UserName = "elias", Email = "elias@test.com" },
+                new User { DisplayName = "Hephaestots", UserName = "vince", Email = "vinc@test.com", IsContentCreator = true },
+                new User { DisplayName = "SilveR", UserName = "elias", Email = "elias@test.com", IsContentCreator = true },
                 new User { DisplayName = "Nabil", UserName = "nabil", Email = "nabil@test.com" },
                 new User { DisplayName = "Venomyox", UserName = "younes", Email = "younes@test.com" }
             };
@@ -206,9 +210,7 @@ namespace Persistence
                         EventTypeId = eventTypes[2].Id,
                         Title = "Past Event 1",
                         Date = DateTime.UtcNow.AddMonths(-2),
-                        Description = "Event 2 months ago",
-                        City = "London",
-                        Venue = "Pub"
+                        Description = "Event 2 months ago"
                     },
                     new Event
                     {
@@ -217,9 +219,7 @@ namespace Persistence
                         EventTypeId = eventTypes[1].Id,
                         Title = "Past Event 2",
                         Date = DateTime.UtcNow.AddMonths(-1),
-                        Description = "Event 1 month ago",
-                        City = "Paris",
-                        Venue = "Louvre"
+                        Description = "Event 1 month ago"
                     },
                     new Event
                     {
@@ -228,9 +228,7 @@ namespace Persistence
                         EventTypeId = eventTypes[0].Id,
                         Title = "Future Event 1",
                         Date = DateTime.UtcNow.AddMonths(1),
-                        Description = "Event 1 month in future",
-                        City = "London",
-                        Venue = "Natural History Museum"
+                        Description = "Event 1 month in future"
                     },
                     new Event
                     {
@@ -239,9 +237,7 @@ namespace Persistence
                         EventTypeId = eventTypes[2].Id,
                         Title = "Future Event 2",
                         Date = DateTime.UtcNow.AddMonths(2),
-                        Description = "Event 2 months in future",
-                        City = "London",
-                        Venue = "O2 Arena"
+                        Description = "Event 2 months in future"
                     },
                     new Event
                     {
@@ -250,9 +246,7 @@ namespace Persistence
                         EventTypeId = eventTypes[1].Id,
                         Title = "Future Event 3",
                         Date = DateTime.UtcNow.AddMonths(3),
-                        Description = "Event 3 months in future",
-                        City = "London",
-                        Venue = "Another pub"
+                        Description = "Event 3 months in future"
                     },
                     new Event
                     {
@@ -261,9 +255,7 @@ namespace Persistence
                         EventTypeId = eventTypes[0].Id,
                         Title = "Future Event 4",
                         Date = DateTime.UtcNow.AddMonths(4),
-                        Description = "Event 4 months in future",
-                        City = "London",
-                        Venue = "Yet another pub"
+                        Description = "Event 4 months in future"
                     },
                     new Event
                     {
@@ -272,9 +264,7 @@ namespace Persistence
                         EventTypeId = eventTypes[2].Id,
                         Title = "Future Event 5",
                         Date = DateTime.UtcNow.AddMonths(5),
-                        Description = "Event 5 months in future",
-                        City = "London",
-                        Venue = "Just another pub"
+                        Description = "Event 5 months in future"
                     },
                     new Event
                     {
@@ -283,9 +273,7 @@ namespace Persistence
                         EventTypeId = eventTypes[1].Id,
                         Title = "Future Event 6",
                         Date = DateTime.UtcNow.AddMonths(6),
-                        Description = "Event 6 months in future",
-                        City = "London",
-                        Venue = "Roundhouse Camden"
+                        Description = "Event 6 months in future"
                     },
                     new Event
                     {
@@ -294,9 +282,7 @@ namespace Persistence
                         EventTypeId = eventTypes[0].Id,
                         Title = "Future Event 7",
                         Date = DateTime.UtcNow.AddMonths(7),
-                        Description = "Event 2 months ago",
-                        City = "London",
-                        Venue = "Somewhere on the Thames"
+                        Description = "Event 2 months ago"
                     },
                     new Event
                     {
@@ -305,11 +291,13 @@ namespace Persistence
                         EventTypeId = eventTypes[2].Id,
                         Title = "Future Event 8",
                         Date = DateTime.UtcNow.AddMonths(8),
-                        Description = "Event 8 months in future",
-                        City = "London",
-                        Venue = "Cinema"
+                        Description = "Event 8 months in future"
                     }
                 };
+
+                events[0].Contributors.Add(new EventContributor { Event = events[0], User = users[1], IsAdmin = true, Status = Domain.Enums.ContributorStatus.Creator });
+                events[0].Contributors.Add(new EventContributor { Event = events[0], User = users[2], Status = Domain.Enums.ContributorStatus.Accepted });
+
                 await _context.Events.AddRangeAsync(events);
                 await _context.SaveChangesAsync();
             }
